requirements.txt
pillow==4.2.1

pytest==3.2.2
pytest-mock==1.6.2
pytest-testmon==0.9.6
pytest-watch==4.1.0
httpretty==0.8.14

flake8==3.4.1
ipython==6.1.0
yapf==0.17.0
isort==4.2.15
